agregar meta data
