+++ src/app/config/metadata.tsx
@@ -1,0 +1,17 @@
+export const metadata = {
+    title: "Mi página increíble",
+    description: "Esta es una descripción optimizada para SEO",
+    openGraph: {
+      title: "Mi página increíble - OpenGraph",
+      description: "Descripción para compartir en redes sociales",
+      url: "https://mi-sitio.com",
+      images: [
+        {
+          url: "/imagen-og.png",
+          width: 1200,
+          height: 630,
+        },
+      ],
+    },
+  };
+
